atlas update - major

+++ CanlabCore/@region/region.m
@@ -377,8 +377,8 @@ classdef region
                 obj(i).dim = mask.volInfo.dim;
                 obj(i).numVox = size(obj(i).XYZ, 2);
                 
-                obj(i).center = center_of_mass(myXYZ, obj(i).Z);
-                obj(i).mm_center = center_of_mass(obj(i).XYZmm, obj(i).Z);
+                obj(i).center = center_of_mass(myXYZ, double(obj(i).Z));
+                obj(i).mm_center = center_of_mass(obj(i).XYZmm, double(obj(i).Z));
                 
                 [dd, ff, ee] = fileparts(mask.volInfo.fname);
                 obj(i).custom_info1 = [ff ee];
